move contentView Container to it's own component
